Hide upgrade/sell buttons when a tower is not selected
