+++ Assets/Scenes/SampleScene.unity
@@ -1805,7 +1805,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 272705727}
-  - component: {fileID: 272705728}
+  - component: {fileID: 272705729}
   m_Layer: 5
   m_Name: TowerButtons
   m_TagString: Untagged
@@ -1836,7 +1836,7 @@ RectTransform:
   m_AnchoredPosition: {x: 205, y: -5}
   m_SizeDelta: {x: 90, y: 90}
   m_Pivot: {x: 0, y: 1}
---- !u!114 &272705728
+--- !u!114 &272705729
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -1845,16 +1845,17 @@ MonoBehaviour:
   m_GameObject: {fileID: 272705726}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 97c4894bd8e7057429af0e033cb8c61d, type: 3}
+  m_Script: {fileID: 11500000, guid: dc0c77e806f9ad549a22dd8b0af002cb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  Buttons:
+  UpgradeButtons:
   - {fileID: 1274034837}
   - {fileID: 1134157955}
-  - {fileID: 278093345}
-  Tooltips:
+  UpgradeTooltips:
   - {fileID: 1274034841}
   - {fileID: 1134157959}
+  SellButtons:
+  - {fileID: 1250217732}
 --- !u!1 &278093343
 GameObject:
   m_ObjectHideFlags: 0
@@ -7057,9 +7058,8 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 272705728}
-        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.UpgradeTowerButtonsManager,
-          Assembly-CSharp
+      - m_Target: {fileID: 272705729}
+        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.TowerButtonsManager, Assembly-CSharp
         m_MethodName: Refresh
         m_Mode: 1
         m_Arguments:
@@ -7098,9 +7098,8 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 272705728}
-        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.UpgradeTowerButtonsManager,
-          Assembly-CSharp
+      - m_Target: {fileID: 272705729}
+        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.TowerButtonsManager, Assembly-CSharp
         m_MethodName: Refresh
         m_Mode: 1
         m_Arguments:
@@ -7139,9 +7138,8 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 272705728}
-        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.UpgradeTowerButtonsManager,
-          Assembly-CSharp
+      - m_Target: {fileID: 272705729}
+        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.TowerButtonsManager, Assembly-CSharp
         m_MethodName: Refresh
         m_Mode: 1
         m_Arguments:
@@ -10442,22 +10440,9 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 272705728}
-        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.UpgradeTowerButtonsManager,
-          Assembly-CSharp
+      - m_Target: {fileID: 272705729}
+        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.TowerButtonsManager, Assembly-CSharp
         m_MethodName: SetTower
-        m_Mode: 0
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
-      - m_Target: {fileID: 1250217732}
-        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.SellTower, Assembly-CSharp
-        m_MethodName: SetState
         m_Mode: 0
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -10482,22 +10467,9 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 272705728}
-        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.UpgradeTowerButtonsManager,
-          Assembly-CSharp
+      - m_Target: {fileID: 272705729}
+        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.TowerButtonsManager, Assembly-CSharp
         m_MethodName: SetTower
-        m_Mode: 0
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
-      - m_Target: {fileID: 1250217732}
-        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.SellTower, Assembly-CSharp
-        m_MethodName: SetState
         m_Mode: 0
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -10558,22 +10530,9 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 272705728}
-        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.UpgradeTowerButtonsManager,
-          Assembly-CSharp
+      - m_Target: {fileID: 272705729}
+        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.TowerButtonsManager, Assembly-CSharp
         m_MethodName: SetTower
-        m_Mode: 0
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
-      - m_Target: {fileID: 1250217732}
-        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.SellTower, Assembly-CSharp
-        m_MethodName: SetState
         m_Mode: 0
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -10694,9 +10653,8 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-      - m_Target: {fileID: 272705728}
-        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.UpgradeTowerButtonsManager,
-          Assembly-CSharp
+      - m_Target: {fileID: 272705729}
+        m_TargetAssemblyTypeName: TDDemo.Assets.Scripts.UI.TowerButtonsManager, Assembly-CSharp
         m_MethodName: Clear
         m_Mode: 1
         m_Arguments:
