+++ Assets/Scripts/UI/TowerButtonsManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 97c4894bd8e7057429af0e033cb8c61d
+guid: dc0c77e806f9ad549a22dd8b0af002cb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
